telegram_bot/bot.py
# bot.py
import os
from telegram.ext import ApplicationBuilder, MessageHandler, filters, ContextTypes
from telegram import Update
from dotenv import load_dotenv

load_dotenv()

BOT_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")

def clean_text(text: str) -> str:
    return text.replace("*", "").replace("#", "")

async def handle_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
    text = update.message.text
    await update.message.reply_text(clean_text(f"You said: {text}"))

async def main():
    app = ApplicationBuilder().token(BOT_TOKEN).build()
    app.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, handle_message))
    print("🤖 Bot is running (polling)...")
    await app.run_polling()

if __name__ == "__main__":
    import asyncio
    asyncio.run(main())
